+++ src/Config.cpp
@@ -20,10 +20,12 @@
 namespace PacBio {
 namespace BAM {
 
-// Initialized to -1 to indicate default. We will set this to HTS_LOG_OFF unless
-// client code overrides. This keeps htslib from polluting stdout/stderr on its own.
-//
-int HtslibVerbosity = -1;
+// Disable htslib's own logging at startup. Client code can still override with
+// hts_set_log_level(HTS_LOG_FOO).
+static const int DisableHtslibLogging = []() {
+    hts_set_log_level(HTS_LOG_OFF);
+    return 0;
+}();
 
 bool DoesHtslibSupportLongCigar()
 {