+++ CHANGELOG.md
@@ -5,8 +5,21 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Active
 
+### Removed
+ - 'HtslibVerbosity' setting in pbbam config. htslib's own logging to stderr is
+   simply disabled at startup by default. In the rare case that client needs to
+   override this behavior, use 'hts_set_log_level()` instead.
+
+## [1.4.0] - 2020-05-22
+
 ### Added
  - SamReader
+
+### Changed
+ - Data::SNR (from pbcopper) values are now float, not double.
+
+### Removed
+ - BamRecordBuilder
 
 ## [1.3.0] - 2020-04-24